Fix `alchemist-buffer--error-link-options` to match compilation errors
This commit changes the word boundaries in the regexp
`alchemist-buffer--error-link-options`. This regexp is used to match
compilation errors in `mix-buffer`.

The previous value of `"^\\([-A-Za-z0-9./_]+\\):\\([0-9]+\\)\\(?:
warning\\)?$"` would not be able to match any of the errors in the
excerpt below:

```
  1) test can add to filter (BloomFilterTest)
     test/bloom_filter_test.exs:29
     ** (throw) :error
     stacktrace:
       (bloom_filter) lib/bloom_filter.ex:18: BloomFilter.test/2
       (elixir) lib/enum.ex:2139: Enum.do_all?/2
       test/bloom_filter_test.exs:35
```

With the new value of `"\\([-A-Za-z0-9./_]+\\):\\([0-9]+\\)\\(?:
warning\\)?"`, we have the following matches:

  - match: lib/bloom_filter.ex:18
    1st capture group: lib/bloom_filter.ex
    2nd capture group: 18

  - match: lib/enum.ex:2139
    1st capture group: lib/enum.ex
    2nd capture group: 2139

  - match: test/bloom_filter_test.exs:35
    1st capture group: test/bloom_filter_test.exs
    2nd capture group: 35

With this change, it is possible to follow the error corresponding
`lib/bloom_filter.ex`.