Make company-mode behavior changes buffer-local
I'm having problems with the modifications injected into company-mode by
alchemist. For example, I use <kbd>C-h</kbd> as `backward-delete-char`
instead of `help` and alchemist rebinds it to
`alchemist-company--show-documentation` which is definitely not what I
want when I'm writing code in other languages.

This commit tries to isolate the modifications provided by alchemist
only to buffers with `alchemist-mode` enabled by making use of
buffer-local variables.

I know very little about `company` and I'm not sure if this is the right
way to address the issue.