Add alchemist-utils--regex-in-buffer-p
This function checks if a regex occurs at least once in a given buffer,
and returns non-nil if it does or nil if it doesn't.

It will be useful in at least two cases: in alchemist-goto.el, when
checking if there are definitions in the current buffer, and in
alchemist-test-mode.el, when checking if there are tests in the current buffer.