Update alchemist-phoenix-project-p
Unfortunately, modern phoenix projects no longer have a "web"
directory, opting instead for a "lib/<project_name>_web"
directory. This is mentioned in passing on the "Up and Running" page
for Phoenix: https://hexdocs.pm/phoenix/up_and_running.html#content.

Since it would be difficult to ascertain the name of the project in
order to check this directory explicitly, I opted instead to check
whether the string ":phoenix" occurs in the "mix.exs" file of the
project. This would be expected in the dependencies section of that
file.

We could, alternately, check for the existence of a "lib/*_web"
directory.